+++ CHANGELOG.md
@@ -4,6 +4,8 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [unreleased]
 
+- Feature: Add `TagList` component.
+- Feature: Add `TagIcon` component.
 - Feature: Create `SupTitle` component.
 - Fix: Delete description.
 - Fix: Redesign new subtitles and add border-bottom.